simula-race/requirements.txt
discord.py>=2.4.0
SQLAlchemy>=2.0.30
alembic>=1.13.1
APScheduler>=3.10.4
Pillow>=10.3.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
pytz>=2024.1
aiosqlite>=0.20.0
rapidfuzz>=3.9.0
greenlet>=3.2.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
